refactor(wallet-sdk,core-ledger-client-types): move wrappedcommand type
